Restricted LS candidate moves
