+++ pyvrp/accept/AcceptanceCriterion.py
@@ -1,0 +1,6 @@
+from typing import Protocol
+
+
+class AcceptanceCriterion(Protocol):
+    def __call__(self, best: float, current: float, candidate: float) -> bool:
+        ...
