+++ pyvrp/accept/__init__.py
@@ -1,1 +1,2 @@
-from .RecordToRecord import RecordToRecord
+from .AcceptanceCriterion import AcceptanceCriterion as AcceptanceCriterion
+from .RecordToRecord import RecordToRecord as RecordToRecord
