+++ pyvrp/search/DestroyRepair.py
@@ -13,11 +13,13 @@ class DestroyRepair:
         self,
         data: ProblemData,
         rng: RandomNumberGenerator,
+        nbhd: list[list[int]],
         destroy_ops: list,
         repair_ops: list,
     ):
         self._data = data
         self._rng = rng
+        self._nbhd = nbhd
         self._destroy_ops = destroy_ops
         self._repair_ops = repair_ops
 
@@ -49,20 +51,10 @@ class DestroyRepair:
             The improved solution. This is not the same object as the
             solution that was passed in.
         """
-        d_idx = self._rng.randint(len(self._destroy_ops))
-        r_idx = self._rng.randint(len(self._repair_ops))
+        rng = self._rng
 
-        destroy_op = self._destroy_ops[d_idx]
-        repair_op = self._repair_ops[r_idx]
+        d_idx = rng.randint(len(self._destroy_ops))
+        d_op = self._destroy_ops[d_idx]
 
-        destroyed = destroy_op(
-            self._data, solution, cost_evaluator, self._rng, neighbours
-        )
-        return repair_op(
-            self._data, destroyed, cost_evaluator, self._rng, neighbours
-        )
-
-    def register(
-        self, current: Solution, perturbed: Solution, candidate: Solution
-    ):
-        pass
+        # NOTE only destroy now, LS handles repair.
+        return d_op(self._data, solution, cost_evaluator, rng, self._nbhd)
